ADD: product controller
